[java-codegen] Add support for size type
Generate code for size type in api file as uint32

Resolves: LE-8167

+++ framework/tools/ifgen/langJava/codeGenHelpers.py
@@ -55,7 +55,7 @@ _BasicTypeMapping = {
     interfaceIR.BOOL_TYPE:   "boolean",
     interfaceIR.CHAR_TYPE:   "byte",
     interfaceIR.DOUBLE_TYPE: "double",
-    interfaceIR.SIZE_TYPE:   "long",
+    interfaceIR.SIZE_TYPE:   "BigInteger",
     interfaceIR.STRING_TYPE: "String",
     interfaceIR.FILE_TYPE:   "FileDescriptor",
     interfaceIR.RESULT_TYPE: "Result",
